requirements.txt
requests==2.24.0
loguru==0.5.3
APScheduler==3.8.0
python_logging_loki